fastboot: enable keypress detection for fastboot entry
For allowing a user to reflash a phone, there is a need to
enter info fastboot mode, without any console command entry.

This can be done now by keeping any keypad key pressed
and booting up the phone. The phone automatically enters
the fastboot mode and waits for usb inputs

Signed-off-by: Vikram Pandita <vikram.pandita@ti.com>