omap4: mmc: deal with card-terminated multi-block writes
In some cases when doing multi-sector writes, mmc_write_data()
will terminate with a transaction complete before we've written
all the data blocks we expected to.

Return the number of blocks written so that omap_mmc_write_sect()
can advance by that count instead of the count it asked to write.

I assume the reason we were seeing this in sparse image writes
is due to the random access patterns defeating the card's buffering.

Signed-off-by: Brian Swetland <swetland@google.com>