usb/udc: add init/exit functions & gadget probe
Those two shall be provided by the selected udc driver and match with
its probe/clean functions. The idea behind it is to init the udc &
gadget only when they are used and not always at board init time.
The gadget probe function is provided by each udc driver and is called
by the gadget for binding.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>